utils: move the file module to utils::fs
