+++ src/crates.rs
@@ -9,7 +9,6 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::str::FromStr;
 use tar::Archive;
-use utils;
 
 const CRATES_ROOT: &str = "https://crates-io.s3-us-west-1.amazonaws.com/crates";
 
@@ -146,7 +145,7 @@ pub fn prepare_crate(krate: &Crate) -> Result<()> {
                 repo.slug(),
                 dir.display()
             );
-            utils::copy_dir(&repo.mirror_dir(), &dir)?;
+            ::utils::fs::copy_dir(&repo.mirror_dir(), &dir)?;
         }
     }
 
@@ -173,7 +172,7 @@ fn dl_registry(name: &str, vers: &str, dir: &Path) -> Result<()> {
     let r = unpack_without_first_dir(&mut tar, dir).chain_err(|| "unable to unpack crate tarball");
 
     if r.is_err() {
-        let _ = utils::remove_dir_all(dir);
+        let _ = ::utils::fs::remove_dir_all(dir);
     }
 
     r
